refactor RoomController and RoomService to use Optional for room retrieval; enhance error handling for room not found scenarios

+++ src/main/java/co/edu/javeriana/lms/booking/controllers/RoomController.java
@@ -1,5 +1,7 @@
 package co.edu.javeriana.lms.booking.controllers;
 
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -75,7 +77,12 @@ public class RoomController {
     @GetMapping("/{id}")
     public ResponseEntity<ApiResponseDto<?>> getRoomById(@PathVariable Long id) {
         log.info("Requesting room with id={}", id);
-        Room room = roomService.findById(id);
+        Optional<Room> room = roomService.findById(id);
+
+        if (room.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(new ApiResponseDto<>(HttpStatus.NOT_FOUND.value(), "Room not found", null, null));
+        }
 
         return ResponseEntity.status(HttpStatus.OK)
                 .body(new ApiResponseDto<>(HttpStatus.OK.value(), "Room found", room, null));
@@ -95,6 +102,14 @@ public class RoomController {
 
         log.info("Requesting deletion of room with id={}", id);
 
+        // Check if room exists
+        Optional<Room> room = roomService.findById(id);
+
+        if (room.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(new ApiResponseDto<>(HttpStatus.NOT_FOUND.value(), "Room not found", null, null));
+        }
+
         roomService.deleteById(id);
 
         return ResponseEntity.status(HttpStatus.OK)
@@ -109,6 +124,14 @@ public class RoomController {
         log.info("Room: id={}, name={}, capacity={}, type={}", room.getId(), room.getName(), room.getCapacity(),
                 room.getType().getName());
 
+        // Check if room exists
+        Optional<Room> roomEntity = roomService.findById(room.getId());
+
+        if (roomEntity.isEmpty()) {
+            log.error("Room not found");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(new ApiResponseDto<>(HttpStatus.NOT_FOUND.value(), "Room not found", null, null));
+        }
 
         // Update room using the save method
         try {
